pyproject.toml
[project]
name = "claude-usage-visualizer"
version = "0.1.0"
description = "Local dashboard to visualize Claude Code usage and estimate costs"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "python-dotenv>=1.0.0",
    "jinja2>=3.0.0",
    "python-multipart>=0.0.6",
    "requests>=2.31.0",
]

[dependency-groups]
dev = [
    "ruff>=0.9.0",
    "pytest>=8.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/app"]

[tool.ruff]
src = ["src"]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]